+++ lib/user_agent/browsers.rb
@@ -4,6 +4,7 @@ require 'user_agent/browsers/base'
 require 'user_agent/browsers/webkit'
 
 require 'user_agent/browsers/acast'
+require 'user_agent/browsers/air_traffic_framework'
 require 'user_agent/browsers/airr'
 require 'user_agent/browsers/alexa'
 require 'user_agent/browsers/amazon_echo'
@@ -13,9 +14,6 @@ require 'user_agent/browsers/android_download_manager'
 require 'user_agent/browsers/apple_core_media'
 require 'user_agent/browsers/apple_home_pod'
 require 'user_agent/browsers/apple_podcasts'
-require 'user_agent/browsers/apple_watch/base'
-require 'user_agent/browsers/apple_watch/atc'
-require 'user_agent/browsers/apple_watch/pod_2_watch'
 require 'user_agent/browsers/audible'
 require 'user_agent/browsers/audio_clip'
 require 'user_agent/browsers/beyond_pod'
@@ -49,6 +47,7 @@ require 'user_agent/browsers/pandora'
 require 'user_agent/browsers/pandora_rss_crawler'
 require 'user_agent/browsers/playstation'
 require 'user_agent/browsers/pocket_casts'
+require 'user_agent/browsers/pod_2_watch'
 require 'user_agent/browsers/pod_mn'
 require 'user_agent/browsers/podbean'
 require 'user_agent/browsers/podcast_addict'
@@ -134,8 +133,8 @@ class UserAgent
       AppleCoreMedia,
       Libavformat,
       OkHttp,
-      AppleWatch::ATC,
-      AppleWatch::Pod2Watch,
+      AirTrafficFramework,
+      Pod2Watch,
       ApplePodcasts,
       AndroidDownloadManager,
       AudioClip,