+++ lib/user_agent/browsers/base.rb
@@ -20,6 +20,7 @@ class UserAgent
       MAC_OS            = 'macOS'
       MOZILLA           = 'Mozilla'
       OSX               = 'OSX'
+      WATCH_OS          = 'watchOS'
       WINDOWS           = 'Windows'
       X11               = 'X11'
 
@@ -33,6 +34,7 @@ class UserAgent
       IPOD_REGEX        = /iPod/.freeze
       MACINTOSH_REGEX   = /Macintosh/.freeze
       OS_X_REGEX        = /OS X/.freeze
+      WATCH_OS_REGEX    = /watchOS/.freeze
       WINDOWS_NT_REGEX  = /Windows NT/.freeze
       WINDOWS_REGEX     = /Windows/.freeze
       X11_REGEX         = /X11/.freeze